refactor(Calcualte): 타입 변경

+++ src/main/java/lotto/domain/Calculate.java
@@ -2,7 +2,7 @@ package lotto.domain;
 
 public class Calculate {
 
-    public double calculateRateOfProfit(int buyAmount, int winnings) {
+    public double calculateRateOfProfit(int buyAmount, long winnings) {
         double division = (double) winnings / buyAmount * 100;
         double round = Math.round(division * 10);
         double profit = round / 10;
